+++ Python/tdw/output_data.py
@@ -727,6 +727,9 @@ class EnvironmentCollision(OutputData):
     def get_contact_point(self, index: int) -> Tuple[float, float, float]:
         return OutputData._get_vector3(self.data.Contacts(index).Point)
 
+    def get_is_floor(self) -> bool:
+        return self.data.Floor()
+
 
 class Volumes(OutputData):
     def get_data(self) -> Vol.Volumes: